Rename our local html -> htmlgen to avoid conflicting with a Py3 builtin.
