+++ mig/shared/freezefunctions.py
@@ -50,7 +50,7 @@ from mig.shared.fileio import checksum_file, write_file, copy_file, copy_rec, \
     makedirs_rec, make_symlink, make_temp_dir, acquire_file_lock, \
     release_file_lock, walk, listdir
 from mig.shared.filemarks import get_filemark, update_filemark
-from mig.shared.html import get_xgi_html_header, get_xgi_html_footer, \
+from mig.shared.htmlgen import get_xgi_html_header, get_xgi_html_footer, \
     man_base_js, themed_styles, themed_scripts, \
     tablesorter_pager
 from mig.shared.pwcrypto import make_path_hash, sorted_hash_algos, default_algo
