+++ mig/shared/functionality/autolaunch.py
@@ -33,7 +33,7 @@ from __future__ import absolute_import
 from mig.shared import returnvalues
 from mig.shared.functional import validate_input_and_cert
 from mig.shared.init import initialize_main_variables
-from mig.shared.html import themed_styles, themed_scripts, menu_items
+from mig.shared.htmlgen import themed_styles, themed_scripts, menu_items
 from mig.shared.settings import load_settings
