+++ mig/shared/functionality/jobman.py
@@ -37,7 +37,7 @@ from mig.shared.base import client_id_dir
 from mig.shared.defaults import default_pager_entries, csrf_backends
 from mig.shared.functional import validate_input_and_cert
 from mig.shared.handlers import get_csrf_limit, make_csrf_token
-from mig.shared.html import themed_styles
+from mig.shared.htmlgen import themed_styles
 from mig.shared.init import initialize_main_variables, find_entry
