+++ mig/shared/functionality/ls.py
@@ -41,7 +41,7 @@ from mig.shared.base import client_id_dir, invisible_path
 from mig.shared.defaults import seafile_ro_dirname, trash_destdir, csrf_field
 from mig.shared.functional import validate_input
 from mig.shared.handlers import get_csrf_limit, make_csrf_token
-from mig.shared.html import fancy_upload_js, fancy_upload_html, confirm_js, \
+from mig.shared.htmlgen import fancy_upload_js, fancy_upload_html, confirm_js, \
     confirm_html, themed_styles
 from mig.shared.init import initialize_main_variables, find_entry
 from mig.shared.parseflags import all, long_list, recursive, file_info
