+++ mig/shared/functionality/peersaction.py
@@ -44,7 +44,7 @@ from mig.shared.defaults import peers_filename, peer_kinds, peers_fields, \
     keyword_auto, csrf_field
 from mig.shared.functional import validate_input, REJECT_UNSET
 from mig.shared.handlers import safe_handler, get_csrf_limit
-from mig.shared.html import html_post_helper
+from mig.shared.htmlgen import html_post_helper
 from mig.shared.init import initialize_main_variables, find_entry
 from mig.shared.notification import send_email
 from mig.shared.serial import load, dump
