+++ mig/shared/functionality/settings.py
@@ -46,7 +46,7 @@ from mig.shared.duplicatikeywords import get_duplicati_specs
 from mig.shared.editing import cm_css, cm_javascript, cm_options, wrap_edit_area
 from mig.shared.functional import validate_input_and_cert
 from mig.shared.handlers import get_csrf_limit, make_csrf_token
-from mig.shared.html import man_base_js, man_base_html, console_log_javascript, \
+from mig.shared.htmlgen import man_base_js, man_base_html, console_log_javascript, \
     twofactor_wizard_html, twofactor_wizard_js, twofactor_token_html, \
     legacy_user_interface, save_settings_js, save_settings_html, menu_items
 from mig.shared.init import initialize_main_variables, find_entry, extract_menu
