+++ mig/shared/functionality/sharelink.py
@@ -38,7 +38,7 @@ from mig.shared.defaults import default_pager_entries, keyword_owners, \
     keyword_members, csrf_field
 from mig.shared.functional import validate_input_and_cert
 from mig.shared.handlers import safe_handler, get_csrf_limit, make_csrf_token
-from mig.shared.html import man_base_js, man_base_html, html_post_helper
+from mig.shared.htmlgen import man_base_js, man_base_html, html_post_helper
 from mig.shared.init import initialize_main_variables, find_entry
 from mig.shared.notification import notify_user_thread
 from mig.shared.pwcrypto import make_hash
