+++ mig/shared/functionality/showfreeze.py
@@ -38,7 +38,7 @@ from mig.shared.freezefunctions import is_frozen_archive, get_frozen_archive, \
     build_freezeitem_object, brief_freeze, pending_archives_update, TARGET_PATH
 from mig.shared.functional import validate_input_and_cert, REJECT_UNSET
 from mig.shared.handlers import safe_handler, get_csrf_limit, make_csrf_token
-from mig.shared.html import man_base_js, man_base_html, html_post_helper
+from mig.shared.htmlgen import man_base_js, man_base_html, html_post_helper
 from mig.shared.init import initialize_main_variables, find_entry
 from mig.shared.pwcrypto import sorted_hash_algos
